requirements.txt
numpy
pandas
fastapi
uvicorn
google-generativeai
Pillow
jinja2
python-dotenv
supabase
pydantic